Realtime-Object-Webapp/fastapi-todo/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
opencv-python==4.9.0.80
numpy==1.26.4
torch==2.2.0
ultralytics==8.1.2
Pillow==10.2.0
python-multipart==0.0.9 